app/requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
typing
pydantic[email]
python-multipart